dtgbot class

+++ src/DecodeTheBot/core/consts.py
@@ -19,4 +19,4 @@ RESTORE_FROM_JSON = os.environ.get("RESTORE_FROM_JSON", "False").lower() == "tru
 BACKUP_JSON = Path(os.environ.get("BACKUP_JSON", BACKUP_DIR / "guru_backup.json"))
 PODCAST_URL = os.environ.get("PODCAST_URL")
 GURU_NAMES_FILE = Path(os.environ.get("GURU_NAMES_FILE", "gurunames.txt"))
-SCRAPER_SLEEP = os.environ.get("SCRAPER_SLEEP", 60 * 10)
+SCRAPER_SLEEP = int(os.environ.get("SCRAPER_SLEEP", 60 * 10))
